Add unsigned int

+++ Mtmchkin.cpp
@@ -124,7 +124,7 @@ bool Mtmchkin::validatePlayerName(std::string &input, std::string &name) {
 bool Mtmchkin::validateClassAndCreatePlayer(std::string input, std::string name, std::unique_ptr<Player> &player) {
     std::string playerClass;
     
-    for(int i = (int)name.size() + 1; i < input.size(); i++) {
+    for(unsigned int i = name.size() + 1; i < input.size(); i++) {
         playerClass[i] = input[i];
     }
